Parent and child Working (TODOs needed and critical error!)
TODO: When a parent is selected, move all childrens with it, and when double clicking an object even inside a parent, only select that object + when an object is selected, show a line covering the object.

Critical error: Find a solution for an error breaking the program. The error takes place when closing the program having a empty game object with a child in it in the scene